breaking(types): move type definitions to src folder
