+++ package.json
@@ -6,7 +6,7 @@
   "author": "Eric Liu (https://github.com/metonym)",
   "svelte": "./src/index.js",
   "main": "./src/index.js",
-  "types": "./types/index.d.ts",
+  "types": "./src/index.d.ts",
   "sideEffects": false,
   "scripts": {
     "dev": "rollup -cw",
@@ -33,8 +33,7 @@
     "input"
   ],
   "files": [
-    "src",
-    "types"
+    "src"
   ],
   "prettier": {
     "printWidth": 120,
