+++ src/Pincode.svelte.d.ts
